refactor: project folder structure | add: access `values` from `useForm`
